linter_is_satisfied
