+++ main/src/main.js
@@ -2,20 +2,19 @@ import preparePage, {
     mathQuillify,
 } from "./js/preparePage.js";
 import {
-    randomId,
-    editorAction,
+    // editorAction,
     sensorTimer
 } from "./js/editor.js";
-import decode from "./js/decode.js";
+import decode, {encode} from "./js/decode.js";
 // import config from "./js/config.json";
 
 // H5Pbridge
 export {
     decode,
-    editorAction,
+    encode,
+    // editorAction,
     mathQuillify,
     preparePage,
-    randomId,
     sensorTimer
 };
